+++ dummy-date/src/ui/pages/AboutPage.tsx
@@ -1,0 +1,7 @@
+export function AboutPage() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen">
+      <h1 className="text-4xl font-bold">About Page</h1>
+    </div>
+  );
+}
