+++ dummy-date/src/ui/pages/index.ts
@@ -1,0 +1,2 @@
+export * from "./AboutPage";
+export * from "./ChangeDatePage";
